requirements.txt
flask
line-bot-sdk
langchain
langchain
langchain-community
langchainhub
langchain-openai
langchain-chroma
langchain-groq